test(favorite): update tests to reflect new favorite logic behaviour
